Refactor delete_admin.php and delete_user.php
Create new files delete_admin.php and delete_user.php in the Backend directory. These files handle the deletion of admin and user records from the database respectively. The code checks if the user is logged in and then deletes the corresponding record based on the provided ID. If the deletion is successful, the user is redirected to the respective page with a success message. Otherwise, an error message is displayed. These changes improve the functionality of the system by allowing admins to delete user records.