Refactor delete_admin.php and delete_user.php
Create new files delete_admin.php and delete_user.php to handle the deletion of admins and users respectively. These files check if the user is logged in, validate the input, and execute the deletion query. If successful, the user is redirected to the respective pages with a success message. Otherwise, an error message is displayed.

These changes improve the backend functionality by providing a dedicated mechanism for deleting admins and users.